Update appearance.py
I added some commands to the event. 

Command: An Avatar command that allows you to put in an image URL converts it into data and then prompts you on the file name. Option 2 is an attachment - this allows you to send an attachment and the file name of that attachment will be added. Also, I made this so only I can use it.